+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.139.2] - 2025-08-04
+### Added
+- Quick Links (TOC) on the Appearance tab of the Settings page.
+- New setting: "Align RTL post bodies and user descriptions to the right"
+  (enabled by default).
+
 ## [1.139.0] - 2025-08-02
 ### Changed
 - The RTL texts of post body and user's self-description are now right-aligned.